Refactor: Project structure - swagger yaml
Moved some of the jsdocs to yaml, as jsdocs are a pain to write as lack of full text editor support.
We should consider moving purely to a non jsdoc route for better text editor support.

Next refactor will be to use the router object to allow the api to be split into files