Added /playback/refresh
Allow API to drive when to refresh the software components.
Per project and or software component implements what to do with this message.
Can be ignored or used.
Further expected use cases are for periods of non use or least impact, run expensive cleanups or other operational or functional requirement sed tasks.